data/pyproject.toml
[project]
name = "data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiofiles>=25.1.0",
    "aiosqlite>=0.22.1",
    "httpx>=0.28.1",
    "numpy>=2.4.3",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "scipy>=1.17.1",
    "tqdm>=4.67.3",
]
